+++ src/app/app.routes.ts
@@ -1,3 +1,10 @@
 import { Routes } from '@angular/router';
+import { SimpleComponent } from './simple/simple.component';
 
-export const routes: Routes = [];
+export const routes: Routes = [
+    {
+        path: 'simple',
+        component: SimpleComponent,
+        pathMatch: 'full'
+    }
+];
