+++ src/app/simple/simple.component.html
@@ -1,0 +1,14 @@
+<div>
+    <input type="string" [(ngModel)]="currentTodo" />
+    <button *ngIf="!updateTodoEnable" (click)="addTodo()">Add Todo</button>
+    <button *ngIf="updateTodoEnable" (click)="updateTodo()">Update Todo</button>
+</div>
+
+<div class="list-of-todo">
+    <div *ngFor="let todo of todoList;let i=index;" >
+        <div>
+            <input type="checkbox" [checked]="todo.status == 'completed'"  (change)="checkTodo(todo)" >
+            <span [ngStyle]="{ 'text-decoration' : todo.status == 'completed' ? 'line-through' : '' }" (click)="clickTodo(todo)">{{i+1}}.{{todo.task}}</span>
+        </div>
+    </div>
+</div>
